# back-end/docs/swagger.json
{
  "openapi": "3.0.0",
  "paths": {
...
    "/api/notifications": {
      "get": {
        "operationId": "NotificationsController_all",
        "summary": "List all notifications",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "post": {
        "operationId": "NotificationsController_create",
        "summary": "Create notification",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/CreateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "201": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/user/{userId}": {
      "get": {
        "operationId": "NotificationsController_byUser",
        "summary": "List notifications by user",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "userId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/{id}": {
      "get": {
        "operationId": "NotificationsController_one",
        "summary": "Get one notification",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "patch": {
        "operationId": "NotificationsController_update",
        "summary": "Update notification",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "requestBody": {
          "required": true,
          "content": {
            "application/json": {
              "schema": {
                "$ref": "#/components/schemas/UpdateNotificationDto"
              }
            }
          }
        },
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      },
      "delete": {
        "operationId": "NotificationsController_remove",
        "summary": "Delete notification",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/notifications/{id}/read": {
      "patch": {
        "operationId": "NotificationsController_read",
        "summary": "Mark notification as read",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "id",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Notifications"
        ]
      }
    },
    "/api/portfolio/{learnerId}": {
      "get": {
        "operationId": "PortfolioController_one",
        "summary": "Get learner portfolio",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Portfolio"
        ]
      }
    },
    "/api/portfolio/{learnerId}/holdings": {
      "get": {
        "operationId": "PortfolioController_holdings",
        "summary": "Get learner holdings",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Portfolio"
        ]
      }
    },
    "/api/portfolio/{learnerId}/performance": {
      "get": {
        "operationId": "PortfolioController_performance",
        "summary": "Get learner performance summary",
        "parameters": [
          {
            "name": "x-role",
            "in": "header",
            "description": "Role for RBAC. Accepted: SUPER_USER, ADMIN, INSTRUCTOR, COURSE_PROVIDER, LEARNER. Lowercase frontend aliases also work: superuser, admin, instructor, provider, learner.",
            "required": true,
            "schema": {
              "type": "string",
              "enum": [
                "SUPER_USER",
                "ADMIN",
                "INSTRUCTOR",
                "COURSE_PROVIDER",
                "LEARNER",
                "superuser",
                "admin",
                "instructor",
                "provider",
                "learner"
              ]
            }
          },
          {
            "name": "x-user-id",
            "in": "header",
            "description": "Optional current user id used by the frontend session. Useful for owner checks and demo traceability.",
            "required": false,
            "schema": {
              "type": "string",
              "example": "u6"
            }
          },
          {
            "name": "learnerId",
            "required": true,
            "in": "path",
            "schema": {
              "type": "string"
            }
          }
        ],
        "responses": {
          "200": {
            "description": ""
          },
          "403": {
            "description": "Forbidden: missing or unauthorized role header"
          }
        },
        "tags": [
          "Portfolio"
        ]
      }
    },
...
  },
  "info": {
    "title": "IndAI Review-4 Paper Trading API",
    "description": "NestJS modular backend with in-memory arrays, RBAC through x-role header, DTO validation, Swagger and minute-stock data endpoints.",
    "version": "1.0",
    "contact": {}
  },
...
  "servers": [],
  "components": {
    "schemas": {
...
      "CreateNotificationDto": {
        "type": "object",
        "properties": {
          "userId": {
            "type": "string",
            "example": "u6"
          },
          "title": {
            "type": "string",
            "example": "New session scheduled"
          },
          "message": {
            "type": "string",
            "example": "New session scheduled"
          },
          "type": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warning",
              "error",
              "trade",
              "course",
              "assignment",
              "session",
              "student"
            ],
            "example": "info"
          },
          "read": {
            "type": "boolean",
            "example": false
          },
          "id": {
            "type": "string",
            "example": "n1"
          },
          "createdAt": {
            "type": "string",
            "example": "2026-05-03T00:00:00.000Z"
          }
        },
        "required": [
          "userId",
          "message"
        ]
      },
      "UpdateNotificationDto": {
        "type": "object",
        "properties": {
          "title": {
            "type": "string"
          },
          "message": {
            "type": "string"
          },
          "type": {
            "type": "string",
            "enum": [
              "info",
              "success",
              "warning",
              "error",
              "trade",
              "course",
              "assignment",
              "session",
              "student"
            ]
          },
          "read": {
            "type": "boolean"
          }
        }
      },
...
    }
  }
}